chore: consolidate semimodule with  module
