+++ algebra/src/semimodule/mod.rs
@@ -1,0 +1,12 @@
+
+use crate::{
+  module::TwoSidedSemimodule,
+  ring::Semiring,
+  vector::VectorSpace,
+};
+
+
+pub trait Semimodule: VectorSpace + Multiplicative
+where <Self as TwoSidedSemimodule>::Semiring: Semiring {
+}
+pub mod tropical;
